src/notification/requirements.txt
aws-lambda-powertools
boto3
google-api-python-client
google-auth
sentry-sdk
